+++ uplink_python/__init__.py
@@ -1,1 +1,1 @@
-""" Python Bindings for Storj (V3) """
+"""Python Bindings for Storj (V3)"""